使用雪花算法生成消息id，类型int64

+++ internal/logic/model/db.go
@@ -38,7 +38,7 @@ type OfflineMessage struct {
 
 //Message message model
 type Message struct {
-	ID        int32    `bson:"_id"`
+	ID        int64    `bson:"_id"`
 	Type      int      `bson:"type"`
 	Seq       int32    `bson:"seq"`
 	Operation int32    `bson:"opration"`
